:art: Made routes modular
Everything is modular now OwO
