+++ views/partials/navbar.ejs
@@ -7,7 +7,7 @@
 				/></a>
 			</li>
 			<li>
-				<a class="noti" href="notifications"
+				<a class="noti" href="/notifications"
 					><img src="/public/icons/notifications.svg" alt=""
 				/></a>
 			</li>
